+++ sdks/java/io/debezium/src/test/java/org/apache/beam/io/debezium/DebeziumIOMySqlConnectorIT.java
@@ -108,8 +108,8 @@ public class DebeziumIOMySqlConnectorIT {
 
   @Test
   public void testDebeziumSchemaTransformMysqlRead() throws InterruptedException {
-    int writeSize = 500;
-    int testTime = writeSize * 200;
+    long writeSize = 500L;
+    long testTime = writeSize * 200L;
     MY_SQL_CONTAINER.start();
 
     PipelineOptions options = PipelineOptionsFactory.create();
@@ -147,7 +147,8 @@ public class DebeziumIOMySqlConnectorIT {
     PCollection<Row> result =
         PCollectionRowTuple.empty(readPipeline)
             .apply(
-                new DebeziumReadSchemaTransformProvider(true, writeSize + 4, testTime)
+                new DebeziumReadSchemaTransformProvider(
+                        true, Long.valueOf(writeSize).intValue() + 4, testTime)
                     .from(
                         DebeziumReadSchemaTransformProvider.DebeziumReadSchemaTransformConfiguration
                             .builder()
@@ -164,7 +165,8 @@ public class DebeziumIOMySqlConnectorIT {
     PAssert.that(result)
         .satisfies(
             rows -> {
-              assertThat(Lists.newArrayList(rows).size(), equalTo(writeSize + 4));
+              assertThat(
+                  Lists.newArrayList(rows).size(), equalTo(Long.valueOf(writeSize + 4).intValue()));
               return null;
             });
     Thread writeThread = new Thread(() -> writePipeline.run().waitUntilFinish());
